+++ x-pack/plugins/transform/public/app/common/request.test.ts
@@ -6,7 +6,7 @@
 
 import { PivotGroupByConfig } from '../common';
 
-import { StepDefineExposedState } from '../sections/create_transform/components/step_define/step_define_form';
+import { StepDefineExposedState } from '../sections/create_transform/components/step_define';
 import { StepDetailsExposedState } from '../sections/create_transform/components/step_details/step_details_form';
 
 import { PIVOT_SUPPORTED_GROUP_BY_AGGS } from './pivot_group_by';
