+++ x-pack/plugins/transform/public/app/sections/create_transform/components/advanced_source_editor/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { AdvancedSourceEditor } from './advanced_source_editor';
