+++ x-pack/plugins/transform/public/app/sections/create_transform/components/switch_modal/switch_modal.tsx
