+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -15481,14 +15481,12 @@
     "xpack.transform.stepDefineForm.advancedSourceEditorApplyButtonText": "変更を適用",
     "xpack.transform.stepDefineForm.advancedSourceEditorAriaLabel": "クエリの詳細エディター",
     "xpack.transform.stepDefineForm.advancedSourceEditorHelpText": "高度なエディターでは、変換のソースクエリ句を編集できます。",
-    "xpack.transform.stepDefineForm.advancedSourceEditorLabel": "ソースクエリ句",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalBodyText": "クエリバーに戻すと、編集内容が失われます。",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalConfirmButtonText": "クエリバーに切り替え",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalTitle": "編集内容は失われます",
     "xpack.transform.stepDefineForm.aggExistsErrorMessage": "「{aggName}」という名前の集約構成は既に存在します。",
     "xpack.transform.stepDefineForm.aggregationsLabel": "アグリゲーション（集計）",
     "xpack.transform.stepDefineForm.aggregationsPlaceholder": "集約を追加…",
-    "xpack.transform.stepDefineForm.formHelp": "変換は、ピボット用のスケーラブルで自動化されたプロセスです。開始するにはグループ分けの条件と集約を少なくとも 1 つ選んでください。",
     "xpack.transform.stepDefineForm.groupByExistsErrorMessage": "「{aggName}」という名前のグループ分け構成は既に存在します。",
     "xpack.transform.stepDefineForm.groupByLabel": "グループ分けの条件",
     "xpack.transform.stepDefineForm.groupByPlaceholder": "グループ分けの条件フィールドを追加…",
@@ -15497,8 +15495,6 @@
     "xpack.transform.stepDefineForm.nestedAggListConflictErrorMessage": "「{aggListName}」とネスティングの矛盾があるため、構成「{aggName}」を追加できませんでした。",
     "xpack.transform.stepDefineForm.nestedConflictErrorMessage": "「{aggNameCheck}」とネスティングの矛盾があるため、構成「{aggName}」を追加できませんでした。",
     "xpack.transform.stepDefineForm.nestedGroupByListConflictErrorMessage": "「{groupByListName}」とネスティングの矛盾があるため、構成「{aggName}」を追加できませんでした。",
-    "xpack.transform.stepDefineForm.queryHelpText": "クエリ文字列でソースデータをフィルタリングしてください (オプション)。",
-    "xpack.transform.stepDefineForm.queryLabel": "クエリ",
     "xpack.transform.stepDefineForm.queryPlaceholderKql": "例: {example}",
     "xpack.transform.stepDefineForm.queryPlaceholderLucene": "例: {example}",
     "xpack.transform.stepDefineForm.savedSearchLabel": "保存検索",
@@ -15588,7 +15584,6 @@
     "xpack.transform.transformsWizard.cloneTransformTitle": "クローン変換",
     "xpack.transform.transformsWizard.createTransformTitle": "変換の作成",
     "xpack.transform.transformsWizard.stepCreateTitle": "作成",
-    "xpack.transform.transformsWizard.stepDefineTitle": "ピボットの定義",
     "xpack.transform.transformsWizard.stepDetailsTitle": "ジョブの詳細",
     "xpack.transform.transformsWizard.transformDocsLinkText": "変換ドキュメント",
     "xpack.transform.wizard.nextStepButton": "次へ",
