+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -15488,14 +15488,12 @@
     "xpack.transform.stepDefineForm.advancedSourceEditorApplyButtonText": "应用更改",
     "xpack.transform.stepDefineForm.advancedSourceEditorAriaLabel": "高级查询编辑器",
     "xpack.transform.stepDefineForm.advancedSourceEditorHelpText": "高级编辑器允许您编辑数据帧转换的源查询子句。",
-    "xpack.transform.stepDefineForm.advancedSourceEditorLabel": "源查询子句",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalBodyText": "切换回到查询栏，您将会丢失编辑。",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalConfirmButtonText": "切换至查询栏",
     "xpack.transform.stepDefineForm.advancedSourceEditorSwitchModalTitle": "编辑将会丢失",
     "xpack.transform.stepDefineForm.aggExistsErrorMessage": "名称为“{aggName}”的聚合配置已存在。",
     "xpack.transform.stepDefineForm.aggregationsLabel": "聚合",
     "xpack.transform.stepDefineForm.aggregationsPlaceholder": "添加聚合……",
-    "xpack.transform.stepDefineForm.formHelp": "转换是用于数据透视的可伸缩和自动化流程。至少选择一个分组依据和聚合，才能开始。",
     "xpack.transform.stepDefineForm.groupByExistsErrorMessage": "名称为“{aggName}”的分组依据配置已存在。",
     "xpack.transform.stepDefineForm.groupByLabel": "分组依据",
     "xpack.transform.stepDefineForm.groupByPlaceholder": "添加分组依据字段……",
@@ -15504,8 +15502,6 @@
     "xpack.transform.stepDefineForm.nestedAggListConflictErrorMessage": "无法添加配置“{aggName}”，因为与“{aggListName}”有嵌套冲突。",
     "xpack.transform.stepDefineForm.nestedConflictErrorMessage": "无法添加配置“{aggName}”，因为与“{aggNameCheck}”有嵌套冲突。",
     "xpack.transform.stepDefineForm.nestedGroupByListConflictErrorMessage": "无法添加配置“{aggName}”，因为与“{groupByListName}”有嵌套冲突。",
-    "xpack.transform.stepDefineForm.queryHelpText": "使用查询字符串筛选源数据（可选）。",
-    "xpack.transform.stepDefineForm.queryLabel": "查询",
     "xpack.transform.stepDefineForm.queryPlaceholderKql": "例如，{example}",
     "xpack.transform.stepDefineForm.queryPlaceholderLucene": "例如，{example}",
     "xpack.transform.stepDefineForm.savedSearchLabel": "已保存搜索",
@@ -15595,7 +15591,6 @@
     "xpack.transform.transformsWizard.cloneTransformTitle": "克隆转换",
     "xpack.transform.transformsWizard.createTransformTitle": "创建转换",
     "xpack.transform.transformsWizard.stepCreateTitle": "创建",
-    "xpack.transform.transformsWizard.stepDefineTitle": "定义透视",
     "xpack.transform.transformsWizard.stepDetailsTitle": "作业详情",
     "xpack.transform.transformsWizard.transformDocsLinkText": "转换文档",
     "xpack.transform.wizard.nextStepButton": "下一个",
