+++ x-pack/test/functional/apps/transform/creation_index_pattern.ts
@@ -58,6 +58,7 @@ export default function({ getService }: FtrProviderContext) {
           return `user-${this.transformId}`;
         },
         expected: {
+          pivotAdvancedEditorValueArr: ['{', '  "group_by": {', '    "category.keyword": {'],
           pivotAdvancedEditorValue: {
             group_by: {
               'category.keyword': {
@@ -117,6 +118,7 @@ export default function({ getService }: FtrProviderContext) {
           return `user-${this.transformId}`;
         },
         expected: {
+          pivotAdvancedEditorValueArr: ['{', '  "group_by": {', '    "geoip.country_iso_code": {'],
           pivotAdvancedEditorValue: {
             group_by: {
               'geoip.country_iso_code': {
@@ -233,7 +235,7 @@ export default function({ getService }: FtrProviderContext) {
         it('displays the advanced configuration', async () => {
           await transform.wizard.enabledAdvancedPivotEditor();
           await transform.wizard.assertAdvancedPivotEditorContent(
-            testData.expected.pivotAdvancedEditorValue
+            testData.expected.pivotAdvancedEditorValueArr
           );
         });
